projects.yaml: Add Seaborn
Seaborn is a statistical data visualization library in Python that is built on top of Matplotlib. It provides a high-level interface for drawing attractive and informative statistical graphics. The wide exposure of Seaborn is evident from its substantial number (>10k) of GitHub stars and its frequent usage in data analysis and visualization tasks within the Python community.