Move Prosody to emeritus status, update metadata
The Prosody repo was never on Github, and this was linking to a third-party
mirror of the repository. Now that the 0ver-related project metadata is not
really going to change anyway, I've switched it to point to the official repo
and manually entered all the relevant keys into projects.yaml.

The data for the keys can be verified at https://prosody.im/doc/release/ and
https://hg.prosody.im/trunk/tags

Farewell 0ver, it was fun while it lasted, but alas, all good things must come
to an end! :'(